requirements.txt
Django==1.9
djangorestframework==3.4.7
gevent==1.2.1
greenlet==0.4.12
wheel==0.24.0
